# ECommerceApplication/src/main/java/com/scem/ecommerce/dto/CartItemRequest.java
package com.scem.ecommerce.dto;


public class CartItemRequest {
	
	private Long productId;
	private int quantity;


	public Long getProductId() { 
		return productId;
		}

	public void setProductId(Long productId) {
		this.productId = productId; 
	}
	public int getQuantity() { 
		return quantity; 
	}
	public void setQuantity(int quantity) {
		this.quantity = quantity; 
	}
	
}