# ECommerceApplication/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Ecommerce API
  description: API documentation for Ecommerce system (admin and user roles)
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api
    description: Local server
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /cart:
    get:
      tags: [Cart]
      summary: Get current user's cart items
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
      # Role: USER
  /cart/add:
    post:
      tags: [Cart]
      summary: Add product to cart for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
      # Role: USER
  /cart/remove/{productId}:
    delete:
      tags: [Cart]
      summary: Remove product from cart for current user
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated cart after removal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
      # Role: USER
  /cart/update:
    put:
      tags: [Cart]
      summary: Update quantity of product in cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
      # Role: USER
  /cart/clear:
    delete:
      tags: [Cart]
      summary: Clear all items in user's cart
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cart cleared successfully
      # Role: USER
...
  /discounts/product/{productId}:
    post:
      tags: [Discount]
      summary: Create discount for product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: {type: integer}
        - in: query
          name: code
          schema: {type: string}
          required: true
        - in: query
          name: value
          schema: {type: number, format: double}
          required: true
        - in: query
          name: isPercentage
          schema: {type: boolean}
          required: true
        - in: query
          name: startDate
          schema: {type: string}
          required: true
        - in: query
          name: endDate
          schema: {type: string}
          required: true
      responses:
        '200':
          description: Discount created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
      # Role: ADMIN

    get:
      tags: [Discount]
      summary: List discounts by product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: List of discounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Discount'
      # Role: USER or ADMIN
...
  /orders:
    get:
      tags: [Order]
      summary: Get all orders (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      # Role: ADMIN

    post:
      tags: [Order]
      summary: Place order and pay (User only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Order placed with payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
      # Role: USER
  /orders/admin/{id}:
    get:
      tags: [Order]
      summary: Get order by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      # Role: ADMIN
  /orders/{id}/status:
    put:
      tags: [Order]
      summary: Update order status (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      # Role: ADMIN
  /orders/my:
    get:
      tags: [Order]
      summary: Get orders of current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      # Role: USER
  /orders/{id}:
    delete:
      tags: [Order]
      summary: Cancel order (User only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '204':
          description: Order canceled successfully
      # Role: USER
  /payments/{paymentId}:
    get:
      tags: [Payment]
      summary: Get payment details
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payment object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/order/{orderId}:
    post:
      tags: [Payment]
      summary: Create payment for order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema: {type: integer}
        - name: method
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PaymentMethod'
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: transactionId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/{paymentId}/status:
    put:
      tags: [Payment]
      summary: Update payment status
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: {type: integer}
        - name: status
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /products:
    post:
      tags: [Product]
      summary: Add new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      # Role: ADMIN

    get:
      tags: [Product]
      summary: Get all products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      # Role: USER or ADMIN
  /products/{id}:
    put:
      tags: [Product]
      summary: Update product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      # Role: ADMIN

    delete:
      tags: [Product]
      summary: Delete product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted
      # Role: ADMIN

    get:
      tags: [Product]
      summary: Get product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      # Role: USER or ADMIN
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        roles:
          type: array
          items:
            type: string
        # Add other user fields as per your entity
      required:
        - email
        - password
    AuthRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
...
    CartItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
          format: double
    Cart:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
...
    Discount:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        value:
          type: number
          format: double
        isPercentage:
          type: boolean
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        userId:
          type: integer
        totalAmount:
          type: number
          format: double
        # Further order fields can be added
    PaymentRequest:
      type: object
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: number
          format: double
        transactionId:
          type: string
      required:
        - method
        - amount
        - transactionId
    OrderResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
        payment:
          $ref: '#/components/schemas/Payment'
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: number
          format: double
        transactionId:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
    PaymentMethod:
      type: string
      enum:
        - CREDIT_CARD
        - DEBIT_CARD
        - PAYPAL
        - STRIPE
        - BANK_TRANSFER
    PaymentStatus:
      type: string
      enum:
        - PENDING
        - COMPLETED
        - FAILED
        - CANCELED
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
...